chore: separate 'git' crate integration tests
